requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.5.0
tqdm>=4.65.0
tensorboard>=2.13.0
pyyaml>=6.0
# Optional: for using pretrained tokenizers (GPT-2, BERT)
# transformers>=4.30.0
# tokenizers>=0.13.0

